Vertex colour UI is now shown for group members When importing a mesh with vertex colours, ensure that the  scene's DMX version is at least 22 so that they can be exported again Fixed some issues that arose in Blender 2.81 Refactored tests Bumped version number
