+++ .gitignore
@@ -2,3 +2,5 @@ __pycache__
 *.suo
 *.pyperf
 *.VC.*
+.vscode
+TestResults
