+++ io_scene_valvesource/__init__.py
@@ -21,7 +21,7 @@
 bl_info = {
 	"name": "Blender Source Tools",
 	"author": "Tom Edwards (translators: Grigory Revzin)",
-	"version": (3, 0, 2),
+	"version": (3, 0, 3),
 	"blender": (2, 80, 0),
 	"category": "Import-Export",
 	"location": "File > Import/Export, Scene properties",
